Add IncludeCmd directive to config parser
Sometimes, UserCmd & PassCmd are not enough.

While it would be possible to call mbsync with a completely generated
config file, that seems mildly inelegant due to requiring a wrapper. On
the downside, making other options' values available to the executed
command would require some kind of "expando" feature.

Complex shell commands are better delegated to scripts, so we do not add
support for multi-line commands or try to avoid the need for multiple
levels of quoting/escaping.

A more conventional name for the directive would be "Eval", but it seems
preferable to be consistent with the existing *Cmd options. An "Include"
directive might be added later.

This patch has been fixed up somewhat by the maintainer.